+++ src/time/rkf.rs
@@ -15,8 +15,8 @@ use crate::{
     errorhandling::Validation,
     // errorhandling::Validation,
     mesh::Mesh,
-    physics::Physics,
     rhs::Rhs,
+    state::Physics,
     update_everything_from_cons,
     NumFlux,
     TimeIntegrationConfig,
